requirements.txt
spotipy~=2.22.1
python-dotenv~=0.21.1